# openapi.yaml
openapi: 3.0.3
info:
  title: SiteDeck PM API
  description: Project management API for EPC contractors. V1 scope.
  version: 1.0.0
  contact:
    name: Site Deck LLC
servers:
  - url: /api/v1
    description: V1 API
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/schedule/activities:
    get:
      summary: List schedule activities
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Activities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleActivity'
    post:
      summary: Create schedule activity
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleActivityInput'
      responses:
        '201':
          description: Activity created
  /projects/{projectId}/schedule/activities/{activityId}:
    get:
      summary: Get activity
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: activityId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Activity details
    patch:
      summary: Update activity
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: activityId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleActivityInput'
      responses:
        '200':
          description: Updated activity
    delete:
      summary: Delete activity
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: activityId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted activity
  /projects/{projectId}/schedule/baselines:
    get:
      summary: List baselines
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Baselines list
    post:
      summary: Create baseline
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, createdBy]
              properties:
                name: { type: string }
                createdBy: { type: string }
      responses:
        '201':
          description: Baseline created
  /projects/{projectId}/schedule/change-requests:
    get:
      summary: List change requests
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Change requests list
    post:
      summary: Create change request
      tags: [Schedule]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [activityId, requestedBy, reasonCode]
              properties:
                activityId: { type: string }
                requestedBy: { type: string }
                reasonCode: { type: string }
                proposedStart: { type: string, format: date-time }
                proposedEnd: { type: string, format: date-time }
                impactDescription: { type: string }
      responses:
        '201':
          description: Change request created
...
  /projects/{projectId}/scope/scope-statements:
    get:
      summary: List scope statements
      tags: [Scope]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Scope statements list
    post:
      summary: Create scope statement
      tags: [Scope]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, createdBy]
              properties:
                content: { type: string }
                createdBy: { type: string }
      responses:
        '201':
          description: Scope statement created
  /projects/{projectId}/scope/change-orders:
    get:
      summary: List change orders
      tags: [Scope]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Change orders list
    post:
      summary: Create change order
      tags: [Scope]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coNumber, date, description]
              properties:
                coNumber: { type: string }
                date: { type: string, format: date-time }
                description: { type: string }
                dollarValue: { type: number }
                scheduleImpact: { type: number }
      responses:
        '201':
          description: Change order created
...
  /projects/{projectId}/risk/risk-items:
    get:
      summary: List risk items
      tags: [Risk]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Risk items list
    post:
      summary: Create risk item
      tags: [Risk]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, category, probability, impact, owner]
              properties:
                description: { type: string }
                category: { type: string }
                probability: { type: string }
                impact: { type: string }
                owner: { type: string }
                mitigationPlan: { type: string }
      responses:
        '201':
          description: Risk item created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ScheduleActivity:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        description: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        duration: { type: integer }
        percentComplete: { type: number }
        status: { type: string }
        isMilestone: { type: boolean }
        isCritical: { type: boolean }
        totalFloat: { type: number }
        freeFloat: { type: number }
    ScheduleActivityInput:
      type: object
      required: [name, startDate, endDate, duration]
      properties:
        name: { type: string }
        description: { type: string }
        wbsItemId: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        duration: { type: integer }
        percentComplete: { type: number }
        status: { type: string }
        isMilestone: { type: boolean }
        predecessors: { type: object }
        successors: { type: object }
...